requirements.txt
absl-py==2.4.0
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
databricks-sdk==0.84.0
docker==7.1.0
fastapi==0.128.1
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
googlemaps==4.10.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
h11==0.16.0
huey==2.6.0
idna==3.11
immutabledict==4.3.1
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
numpy==2.4.2
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
ortools==9.15.6755
packaging==25.0
pandas==2.3.3
pillow==12.1.0
prettytable==3.17.0
protobuf==6.33.5
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
waitress==3.0.2
wcwidth==0.5.3
Werkzeug==3.1.5
xgboost==3.1.3
zipp==3.23.0

